refactor: 유틸리티 함수 파일명 정리
