+++ src/components/purchase/ChargeInputForm.js
@@ -1,5 +1,5 @@
 import Component from '../../core/Component.js';
-import { $ } from '../../utils/element-utils.js';
+import { $ } from '../../utils/element-tools.js';
 import { SELECTOR } from '../../constants/selector.js';
 
 export default class ChargeInputForm extends Component {
