+++ src/components/table/ReturnTable.js
@@ -1,4 +1,4 @@
-import { createElement, combineElement } from '../../utils/element-utils.js';
+import { createElement, combineElement } from '../../utils/element-tools.js';
 import TableCreate from '../../core/Table.js';
 
 export default class ReturnTable extends TableCreate {
