+++ src/core/State.js
@@ -1,4 +1,4 @@
-import { storageManager } from '../utils/data-utils.js';
+import { storageManager } from '../utils/data-tools.js';
 
 export default class State {
   constructor(value, storageKey = '') {
