+++ src/index.js
@@ -1,5 +1,5 @@
 import { SELECTOR } from './constants/selector.js';
-import { $ } from './utils/element-utils.js';
+import { $ } from './utils/element-tools.js';
 
 import App from './App.js';
