+++ src/models/UserInputCheck.js
@@ -2,7 +2,7 @@ import {
   isEmpty,
   isPositiveNumber,
   isMinimumValue,
-} from '../utils/validator.js';
+} from '../utils/validation-tools.js';
 import {
   DONE,
   ERROR,
